change trait from trigonometric to float math ops, add more methods
